requirements.txt
flask
pillow
tqdm
numpy
opencv-python-headless
imutils
requests
beautifulsoup4
qreader